# docs/swagger.yaml
basePath: /
definitions:
  app.loginReq:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  app.loginResp:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      token_type:
        type: string
    type: object
  app.registerResp:
    properties:
      ok:
        type: boolean
    type: object
...
  ds.User:
    properties:
      email:
        type: string
      id:
        type: integer
      login:
        type: string
      name:
        type: string
      password:
        type: string
      phoneNumber:
        type: string
      role:
        $ref: '#/definitions/role.Role'
    type: object
  role.Role:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-comments:
      Admin: "2"
      Buyer: "0"
      Manager: "1"
    x-enum-varnames:
    - Buyer
    - Manager
    - Admin
host: localhost:8080
info:
  contact: {}
  description: Consultation app
  title: IT Services
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login
      operationId: login
      parameters:
      - description: login info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/app.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.loginResp'
      summary: Login
      tags:
      - auth
  /auth/logout:
    get:
      description: Logout
      operationId: logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Logout
      tags:
      - auth
  /auth/registration:
    post:
      consumes:
      - application/json
      description: Registration
      operationId: registration
      parameters:
      - description: user info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ds.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.registerResp'
      summary: Registration
      tags:
      - auth
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"